pyproject.toml
[tool.poetry]
name = "cardano-stakepool-analytics"
version = "0.1.0"
description = ""
authors = ["prateek-gurjar"]

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.0.1"
pylint = "^2.12.2"
dbt-core = "^1.0.3"
dbt-postgres = "^1.0.3"
sqlfluff = "^0.11.0"
psycopg2 = "^2.9.3"
python-dotenv = "^0.20.0"
pandas = "^1.4.1"
boto3 = "^1.21.38"
pytest-mock = "^3.7.0"
typer = "^0.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
